Add files via upload
Hello repository people... this is a basic full draft of the game, not pretty on some pages but it should still work.
Instructions to run the game (Remember to save them all in one location):
1.Compile Hero.java
2.Compile Assault.java
3.Compile Assassin.java
4.Compile Healer.java
5.Compile Support.java
6.Compile Tank.java
7.Compile Player.java
8.Compile Computer.java
9.Compile gui4.java
10.Compile gui2.java
11. RUN gui2
Please play around with it a bit and let me know if there's anything I NEED to fix (except display of certain pages that don't look pretty and HP not being displayed when selecting a character to heal. I'm also aware the winning condition is 1 less round than the number of rounds you choose to play. I will also comment the code in due course